view photo action

+++ controllers/PhotoController.php
@@ -82,25 +82,6 @@ class PhotoController extends Controller
     }
 
     /**
-     * Updates an existing Photo model.
-     * If update is successful, the browser will be redirected to the 'view' page.
-     * @param integer $id
-     * @return mixed
-     */
-    public function actionUpdate($id)
-    {
-        $model = $this->findModel($id);
-
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
-        }
-    }
-
-    /**
      * Deletes an existing Photo model.
      * If deletion is successful, the browser will be redirected to the 'index' page.
      * @param integer $id
@@ -108,6 +89,10 @@ class PhotoController extends Controller
      */
     public function actionDelete($id)
     {
+        if(!Yii::$app->user->identity) {
+            return $this->redirect(['auth/default/login']);
+        }
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
